Removed config-server, service-registry.

+++ generators/controller/templates/app/src/main/resources/db/migration/flyway/V1__new_table_no_seq.sql
@@ -1,5 +1,5 @@
 create table <%= tableName %> (
-    id int not null auto_increment,
+    id bigint not null auto_increment,
     text varchar(1024) not null,
     primary key (id)
 );
